% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo.R
\name{hugo_md}
\alias{hugo_md}
\title{MD file for hugo websites}
\usage{
hugo_md(
  variant = "commonmark",
  preserve_yaml = TRUE,
  toc = FALSE,
  toc_depth = 3,
  number_sections = FALSE,
  fig_width = 7,
  fig_height = 5,
  fig_retina = 2,
  dev = "png",
  df_print = "default",
  includes = NULL,
  md_extensions = NULL,
  pandoc_args = NULL,
  ext = ".md"
)
}
\arguments{
\item{variant}{"commonmark" instead of "markdown_strict"}

\item{preserve_yaml}{"TRUE" instead of "FALSE"}

\item{toc, toc_depth, number_sections}{See `?rmarkdown::md_document`}

\item{fig_width, fig_height, fig_retina, dev, df_print, includes}{See `?rmarkdown::md_document`}

\item{includes, md_extensions, pandoc_args, ext}{See `?rmarkdown::md_document`}
}
\description{
This format generates a standard markdown file.
This function is simply rmarkdown::md_document with an extra
line to insert `{{< url >}}`.
}
\details{
See `?rmarkdown::md_document` for help on parameters. Note, that
`preserve_yaml` has been set to `TRUE` we need to keep the yaml in the md
document for hugo.
}
