% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo.R
\name{set_knitr_chunk_options}
\alias{set_knitr_chunk_options}
\title{Set default chunk options}
\usage{
set_knitr_chunk_options(
  echo = TRUE,
  eval = TRUE,
  collapse = TRUE,
  cache = TRUE,
  comment = "#>",
  fig.path = ".",
  fig.align = "center",
  fig.retina = 2,
  fig.asp = 0.7,
  fig.width = 500/72,
  dev.args = list(png = list(type = "cairo-png")),
  dev = "svg",
  optipng = "-o1 -quiet",
  ...
)
}
\arguments{
\item{echo}{TRUE}

\item{eval}{TRUE}

\item{collapse}{TRUE}

\item{cache}{TRUE}

\item{comment}{"#>"}

\item{fig.path}{"."}

\item{fig.align}{"center"}

\item{fig.retina}{2}

\item{fig.asp}{0.7}

\item{fig.width}{500 / 72}

\item{dev.args}{list(png = list(type = "cairo-png"))}

\item{dev}{"svg"}

\item{optipng}{"-o1 -quiet"}

\item{...}{Other arguments passed to opts_chunk$set}
}
\description{
Set of standard {knitr} chunk options
}
