% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo.R
\name{set_knitr_hooks}
\alias{set_knitr_hooks}
\title{Set knitr hooks}
\usage{
set_knitr_hooks(optipng = knitr::hook_optipng, ...)
}
\arguments{
\item{optipng}{Use optipng by default}

\item{...}{Other arguments passed to knit_hooks$set}
}
\description{
Set standard {knitr} hooks
}
